requirements.txt
psycopg2==2.6
django==1.7.7
swampdragon
djangorestframework==3.1.1
